Adds the get-files functions
Using cffi to use stat(2), as Lisp itself has no way to get all the
informations like a file mode (which holds permission info).

Going through a `file` class to hold all the necessary information to
write a file in a tar archive.

Note: (truename path) is used to avoid the issue where the pathname may
or may not have the trailing slash. The truename always has it if it's a
folder.